fix: test_docker -> test_python_virtualenv
